requirements.txt

Flask==2.1.1
Flask-PyMongo==2.3.0
Flask-WTF==1.0.1
Flask-Login==0.6.2
Werkzeug==2.1.1
WTForms==3.0.1
